HashDigester version 6.0.0.1 - release (added physicalCrytpoRandomStream support)

+++ HashDigester/hashSuite.h
@@ -1,8 +1,8 @@
 //
 // Creator:    http://www.dicelocksecurity.com
-// Version:    vers.5.0.0.1
+// Version:    vers.6.0.0.1
 //
-// Copyright © 2009-2011 DiceLock Security, LLC. All rights reserved.
+// Copyright © 2009-2012 DiceLock Security, LLC. All rights reserved.
 //
 //                               DISCLAIMER
 //
@@ -44,9 +44,9 @@ namespace DiceLockSecurity {
 			/// Points the first hash algorithm in the suite
 			static const	Hashes firstHash;
 
-			BaseHash*				suite[NumberOfHashes];
-			bool					selfCreatedHash[NumberOfHashes];
-			int						instantiatedHashes;
+			BaseHash*			suite[NumberOfHashes];
+			bool				selfCreatedHash[NumberOfHashes];
+			unsigned short int	instantiatedHashes;
 
 		public:
